refactor: create common filter component (#98016)
